贴纸保存和个人中心的点赞、信息添加时间功能

+++ app/src/main/java/com/flyingeffects/com/ui/model/CreationTemplatePreviewModel.java
@@ -62,7 +62,7 @@ public class CreationTemplatePreviewModel {
     //每100毫秒更新一次进度指针
     private static final int updateCursorIntervalMs = 100;
     private float videoRatio = 1;
-    private int realityDuration;
+    private long realityDuration;
 
     public CreationTemplatePreviewModel(Context context, CreationTemplatePreviewMvpCallback callback, String videoPath) {
         this.mContext = context;
